requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
langchain==0.3.7
langchain-community==0.3.7
langgraph==0.2.56
langchain-openai==0.2.7
langchain-text-splitters==0.3.2
supabase==2.5.0
python-dotenv==1.0.0
pypdf==4.0.0
requests==2.32.0
numpy==1.26.4
python-docx==1.1.2
sse-starlette==2.1.3
pydantic>=2.7.0
# fastembed removed